Move the entire config folder to TS
