+++ packages/@coorpacademy-progression-engine/src/config/index.ts
@@ -1,6 +1,6 @@
-// @flow
-import find from 'lodash/fp/find';
-import type {Config, Engine, Progression} from '../types';
+import {find} from 'lodash';
+
+import {Config, Engine, Progression} from '../types';
 import microlearning from './microlearning';
 import learner from './learner';
 
@@ -11,9 +11,11 @@ const engineConfigurations = {
 
 export const getConfig = (engine: Engine): Config => {
   const engineConfiguration = engineConfigurations[engine.ref];
+
   if (!engineConfiguration) {
     throw new Error(`Unknown engine ${engine.ref}`);
   }
+
   return (
     find({version: engine.version}, engineConfiguration.configurations) ||
     engineConfiguration.defaultConfiguration
