+++ packages/@coorpacademy-progression-engine/src/config/test/get-config.ts
@@ -1,26 +1,13 @@
-// @flow
 import test from 'ava';
-import {getConfigForProgression} from '..';
-import type {Progression} from '../../types';
 
-const progression: Progression = {
-  engine: {ref: 'microlearning', version: '1'},
-  engineOptions: {
-    version: '1'
-  },
-  content: {type: 'chapter', ref: '1.A1'},
-  actions: []
-};
+import {getConfig} from '..';
 
 test('should throw an error if the engine ref is unknown', t => {
-  t.throws(
-    () => getConfigForProgression({...progression, engine: {ref: 'foobar', version: '1'}}),
-    'Unknown engine foobar'
-  );
+  t.throws(() => getConfig({ref: 'foobar', version: '1'}), 'Unknown engine foobar');
 });
 
 test('should return the configuration with the given version if it exists', t => {
-  t.deepEqual(getConfigForProgression(progression), {
+  t.deepEqual(getConfig({ref: 'microlearning', version: '1'}), {
     version: '1',
     lives: 1,
     livesDisabled: false,
@@ -32,7 +19,7 @@ test('should return the configuration with the given version if it exists', t =>
     starsPerResourceViewed: 4,
     remainingLifeRequests: 1
   });
-  t.deepEqual(getConfigForProgression({...progression, engine: {ref: 'learner', version: '1'}}), {
+  t.deepEqual(getConfig({ref: 'learner', version: '1'}), {
     version: '1',
     lives: 3,
     livesDisabled: false,
@@ -46,13 +33,24 @@ test('should return the configuration with the given version if it exists', t =>
   });
 });
 
-test('should merge the engineOptions values from the progression into the resulting configuration', t => {
-  const engineOptions = {livesDisabled: true, maxTypos: 100, version: '1'};
-  t.deepEqual(getConfigForProgression({...progression, engineOptions}), {
+test('should return the default configuration if the engine does not have the given version', t => {
+  t.deepEqual(getConfig({ref: 'microlearning', version: 'foobar'}), {
     version: '1',
     lives: 1,
-    livesDisabled: true,
-    maxTypos: 100,
+    livesDisabled: false,
+    maxTypos: 2,
+    slidesToComplete: 4,
+    answerBoundaryLimit: 5,
+    starsPerAskingClue: -1,
+    starsPerCorrectAnswer: 4,
+    starsPerResourceViewed: 4,
+    remainingLifeRequests: 1
+  });
+  t.deepEqual(getConfig({ref: 'learner', version: 'foobar'}), {
+    version: '1',
+    lives: 3,
+    livesDisabled: false,
+    maxTypos: 2,
     slidesToComplete: 4,
     answerBoundaryLimit: 5,
     starsPerAskingClue: -1,
