+++ packages/@coorpacademy-progression-engine/src/config/test/learner.ts
@@ -1,8 +1,9 @@
 import test from 'ava';
-import uniqBy from 'lodash/fp/uniqBy';
-import microlearning from '../microlearning';
+import {uniqBy} from 'lodash';
 
-const {configurations} = microlearning;
+import learner from '../learner';
+
+const {configurations} = learner;
 
 test("Configurations shouldn't have conflicting version numbers", t => {
   t.is(configurations.length, uniqBy('version', configurations).length);
