+++ packages/@coorpacademy-progression-engine/src/config/test/microlearning.ts
@@ -1,8 +1,9 @@
 import test from 'ava';
-import uniqBy from 'lodash/fp/uniqBy';
-import learner from '../learner';
+import {uniqBy} from 'lodash';
 
-const {configurations} = learner;
+import microlearning from '../microlearning';
+
+const {configurations} = microlearning;
 
 test("Configurations shouldn't have conflicting version numbers", t => {
   t.is(configurations.length, uniqBy('version', configurations).length);
